capstone/requirements.txt
numpy==1.21.6
scipy==1.7.3
librosa==0.9.2
soundfile==0.10.3.post1
pyaudio==0.2.13
noisereduce==2.0.1
python-sounddevice==0.4.6
webrtcvad==2.0.10
openai==1.0.0
python-dotenv==0.19.2
PyQt6==6.4.2
setuptools>=65.5.1
wheel>=0.38.4
rnnoise-python==0.0.1 